add player screen with chapter markers implementation
